pyproject.toml
[tool.poetry]
name = "atm-location-assessment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.17"
geopandas = "^0.13.2"
matplotlib = "^3.7.1"
folium = "^0.14.0"
mapclassify = "^2.5.0"
minio = "^7.1.15"
python-dotenv = "^1.0.0"
streamlit = "^1.24.1"
streamlit-folium = "^0.13.0"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.3"
seaborn = "^0.12.2"
black = "^23.3.0"
prefect = "^2.10.21"
pyrosm = "^0.6.1"
numpy = "<1.24"
shapely = "1.8.5"
h3pandas = "^0.2.4"


[tool.poetry.group.flow]
optional = true

[tool.poetry.group.flow.dependencies]
anyio = "3.7.1"
prefect = "^2.10.21"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"